Updated resume / skills / projects
Outdated packages are wreaking havoc on the build process, so I had to dig into the bundled app.js to add the new info.  I'm going to rebuild the site in React using either create-react-app or gatsby so that issues like this don't pop back up, and also to more accurately reflect my skill set.